Customize Juice theme header and navigation
The default Juice header includes the full navigation menu in the
header, which is already present in the sidebar. This commit
customizes the header to only include a logo and a simplified
menu with "Get Started", "Tips", and "GitHub" links.

Also removes `juice_exclude_menu` since we are manually defining
the header.